Home Made/Projects/email-agent-project/backend/requirements.txt
flask
langchain
langchain-groq
langchain-google-genai
pydantic
python-dotenv
flask
